requirements.txt
asgiref==3.7.2
Babel==2.12.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
customtkinter==5.2.0
darkdetect==0.8.0
decorator==5.1.1
Django==4.2.4
fonttools==4.42.0
future==0.18.3
geocoder==1.38.1
idna==3.4
mutagen==1.46.0
Pillow==10.0.0
pycparser==2.21
pygame==2.5.0
pyglet==2.0.9
pyperclip==1.8.2
pywin32==306
ratelim==0.1.6
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tkcalendar==1.6.1
tkintermapview==1.29
ttkthemes==3.2.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wikipedia==1.4.0
